Refactor: Decompose monolithic routes.ts into modular routing and services
